requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
customtkinter==5.2.2
darkdetect==0.8.0
Eel==0.16.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyparsing==3.1.1
pywin32-ctypes==0.2.2
setuptools==69.0.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.1
